Weather Injector and Weather Injector Test Changes

+++ src/main/java/com/sparta/weatherapiproject/jackson/Rain.java
@@ -5,24 +5,24 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class Rain{
 
 	@JsonProperty("3h")
-	private int jsonMember3h;
+	private Double jsonMember3h;
 
 	@JsonProperty("1h")
-	private int jsonMember1h;
+	private Double jsonMember1h;
 
-	public void setJsonMember3h(int jsonMember3h){
+	public void setJsonMember3h(double jsonMember3h){
 		this.jsonMember3h = jsonMember3h;
 	}
 
-	public int getJsonMember3h(){
+	public Double getJsonMember3h(){
 		return jsonMember3h;
 	}
 
-	public void setJsonMember1h(int jsonMember1h){
+	public void setJsonMember1h(double jsonMember1h){
 		this.jsonMember1h = jsonMember1h;
 	}
 
-	public int getJsonMember1h(){
+	public Double getJsonMember1h(){
 		return jsonMember1h;
 	}
